+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- [2025-06-30] [Add Biome and format](https://github.com/rubriclab/state/commit/ae42f3b8ead59e51ee9e4e83230298e0d41d8ecc)
 - [2025-06-25] [Use package post-commit hooks](https://github.com/rubriclab/state/commit/ed55a495c1bcc561a5f3b9831402e7245385fba9)
 - [2025-06-25] [Add log](https://github.com/rubriclab/state/commit/9e126c6b77536f2f41662974732575899c91ff34)
 - [2025-06-25] [Add optional persistence in Redis](https://github.com/rubriclab/state/commit/2dbb2d72694ac95b2d75adda9650a91f950abeca)